Final (2)

+++ OperatorClass.cs
@@ -22,14 +22,14 @@ namespace Standard_Calculator
 
         public void Add()
         {
-            secondValue = float.Parse(num2).ToString();
-            num2 = (float.Parse(firstValue) + float.Parse(secondValue)).ToString();
+            secondValue = Double.Parse(num2).ToString();
+            num2 = (Double.Parse(firstValue) + Double.Parse(secondValue)).ToString();
         }
 
         public void Subtract()
         {
-            secondValue = float.Parse(num2).ToString();
-            num2 = (float.Parse(firstValue) - float.Parse(secondValue)).ToString();
+            secondValue = Double.Parse(num2).ToString();
+            num2 = (Double.Parse(firstValue) - Double.Parse(secondValue)).ToString();
         }
     }
 }
