+++ OtherOperators.cs
@@ -10,14 +10,14 @@ namespace Standard_Calculator
     {
         public void Multiply()
         {
-            SecondValue = float.Parse(Num2).ToString();
-            Num2 = (float.Parse(FirstValue) * float.Parse(SecondValue)).ToString();
+            SecondValue = Double.Parse(Num2).ToString();
+            Num2 = (Double.Parse(FirstValue) * Double.Parse(SecondValue)).ToString();
         }
 
         public void Divide()
         {
-            SecondValue = float.Parse(Num2).ToString();
-            Num2 = (float.Parse(FirstValue) / float.Parse(SecondValue)).ToString();
+            SecondValue = Double.Parse(Num2).ToString();
+            Num2 = (Double.Parse(FirstValue) / Double.Parse(SecondValue)).ToString();
         }
     }
 }
